% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_ideal.R
\name{gpu_em_ideal}
\alias{gpu_em_ideal}
\title{Run the EM ideal estimator on the GPU}
\usage{
gpu_em_ideal(rcdata, steps, burnin = 0, thin = 1, x = NULL,
  abprior = matrix(c(20, 0, 0, 20), 2, 2), xprior = 1, blocks = 0,
  threads = 0)
}
\arguments{
\item{rcdata}{Roll call object}

\item{steps}{Number of MCMC step}

\item{thin}{Keep every `thin` step}

\item{x}{Start values for ideal points}

\item{abprior}{Prior variance of bill parameters}

\item{xprior}{Prior variance of ideal points}

\item{blocks}{Number of GPU blocks}

\item{threads}{Number of GPU threads}
}
\value{
MCMC object containing posterior draws
}
\description{
Run the EM ideal estimator on the GPU
}
