% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal.R
\name{gpuideal}
\alias{gpuideal}
\title{Run the gpuideal sampler}
\usage{
gpuideal(rcdata, samples, burnin = 0, thin = 1, x = NULL,
  abprior = matrix(c(20, 0, 0, 20), 2, 2), xprior = 1, blocks = 0,
  threads = 0)
}
\arguments{
\item{rcdata}{Roll call object}

\item{samples}{Number of MCMC samples (after burnin)}

\item{burnin}{Number of burn in samples}

\item{thin}{Keep every `thin` sample}

\item{x}{Start values for ideal points}

\item{abprior}{Prior variance of bill parameters}

\item{xprior}{Prior variance of ideal points}

\item{blocks}{Number of GPU blocks}

\item{threads}{Number of GPU threads}
}
\value{
MCMC object containing posterior draws
}
\description{
Run the gpuideal sampler
}
