% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal.R
\name{rc_recode}
\alias{rc_recode}
\title{Recode votes and delete non-votes}
\usage{
rc_recode(dat, codes)
}
\arguments{
\item{dat}{roll call object dataset}

\item{codes}{yea and nay values}
}
\description{
Recode votes and delete non-votes
}
