% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal.R
\name{rescaleIdeal}
\alias{rescaleIdeal}
\title{Rescale posterior draws}
\usage{
rescaleIdeal(mcmcres, dir)
}
\arguments{
\item{mcmcres}{MCMC object containing draws from IRT model}

\item{dir}{A vector that should be positively correlated with the xs -- this sets the direction}
}
\value{
rescaled MCMC object
}
\description{
Rescale posterior draws so that the ideal pt distributons have mean 0 and std dev 1 in 
each iteration.
}
