% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ideal.R
\name{simrc2d}
\alias{simrc2d}
\title{Create a 2-D MC data set for testing}
\usage{
simrc2d(nrc = 100, nmem = 100)
}
\arguments{
\item{nrc}{Number of roll calls}

\item{nmem}{Number of legislators}
}
\value{
rollcall dataset.
}
\description{
Generates simulated roll matrix for testing.
}
