% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ideal.R
\name{test_ideal}
\alias{test_ideal}
\title{Run a test of the gpuideal point routine}
\usage{
test_ideal(nrc = 100, nmem = 100, ...)
}
\arguments{
\item{nrc}{Number of roll calls}

\item{nmem}{Number of legislators}

\item{...}{Additional arguments to `gpuideal`}
}
\value{
MCMC object containing posterior draws
}
\description{
Run a test of the CUDA gpu-based ideal point routine using simulated data.
}
