% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_ideal2d.R
\name{gpu_em_ideal2d}
\alias{gpu_em_ideal2d}
\title{Run the EM ideal estimator on the GPU}
\usage{
gpu_em_ideal2d(rcdata, steps, burnin = 0, thin = 1, x = NULL,
  abprior = diag(rep(25, 3)), xprior = diag(2), blocks = 0,
  threads = 0)
}
\arguments{
\item{rcdata}{Roll call object}

\item{steps}{Number of MCMC step}

\item{thin}{Keep every `thin` step}

\item{x}{Start values for ideal points}

\item{abprior}{Prior variance of bill parameters}

\item{xprior}{Prior variance of ideal points}

\item{blocks}{Number of GPU blocks}

\item{threads}{Number of GPU threads}
}
\value{
MCMC object containing posterior draws
}
\description{
Run the EM ideal estimator on the GPU
}
